backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
passlib[bcrypt]
python-jose
email-validator
reportlab
